ecope-backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.22
alembic>=1.12.0
pymysql>=1.1.0

# Authentication
python-jose[cryptography]>=3.3.0


# ML/AI
transformers>=4.34.1
torch>=2.0.1
langchain>=0.0.335
langgraph
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
sentence-transformers>=2.2.2

# Data Analysis
pandas>=2.1.1
numpy>=1.26.0
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.14.0
wordcloud>=1.9.2

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Utils
python-dotenv>=1.0.0
